split logic into two rules "mechanics" and "circular-deps"
